Refactor components to have functional and pure components
